Build an AnnotationFilter::AnnotationFilter for selecting transcripts

This function helps create an AnnotationFilter::AnnotationFilterList that will combine multiple standard filters based on the flags.

Usage

build_annotation_filter(
  standard.chroms = TRUE,
  remove.na.rows = TRUE,
  protein.coding = TRUE,
  transcript.support = NULL,
  canonical = FALSE,
  gene_id = NULL,
  tx_id = NULL,
  symbol = NULL,
  entrez_id = NULL,
  add.filter = NULL
)

Arguments

standard.chroms

Keep only standard chromosomes. Boolean.

remove.na.rows

Remove transcripts with NA in the ID column. Boolean.

protein.coding

Keep only protein coding genes. Boolean.

transcript.support

Filter by transcript support level. Integer 1-4

canonical

Keep only the ENSEMBL canonical transcript. Boolean.

gene_id

Keep specific ENSEMBL ID(s). Input is a vector

tx_id

Keep specific tx_id(s). Input is a vector

symbol

Keep specific gene symbol(s). Input is a vector

entrez_id

Keep specific ENTREZ ID(s). Input is a vector

add.filter

Optional AnnotationFilter::AnnotationFilterList to include

Value

An AnnotationFilter::AnnotationFilterList object

Examples

# Default: standard chromosomes, removes na tx_ids, and keep protein coding
afl = SeedMatchR::build_annotation_filter()
#> Build AnnotationFilter for transcript features based on the following parameters: 
#> Keep only standard chroms: TRUE
#> Remove rows with NA in transcript ID: TRUE
#> Keep only protein coding genes and transcripts: TRUE
#> Filtering for transcripts with support level: FALSE
#> Keep only the ENSEMBL canonical transcript: FALSE
#> Filtering for specific genes: FALSE
#> Filtering for specific transcripts: FALSE
#> Filtering for specific gene symbols: FALSE
#> Filtering for specific entrez id: FALSE

# Keep only transcripts with support level 1
afl = SeedMatchR::build_annotation_filter(transcript.support = 1)
#> Build AnnotationFilter for transcript features based on the following parameters: 
#> Keep only standard chroms: TRUE
#> Remove rows with NA in transcript ID: TRUE
#> Keep only protein coding genes and transcripts: TRUE
#> Filtering for transcripts with support level: 1
#> Keep only the ENSEMBL canonical transcript: FALSE
#> Filtering for specific genes: FALSE
#> Filtering for specific transcripts: FALSE
#> Filtering for specific gene symbols: FALSE
#> Filtering for specific entrez id: FALSE
#> Warning: replacing previous import ‘S4Arrays::makeNindexFromArrayViewport’ by ‘DelayedArray::makeNindexFromArrayViewport’ when loading ‘SummarizedExperiment’