Search input Biostrings::DNAStringSet for a set of defined sequences

This function will use Biostrings::vcountpattern to find matches to the input siRNA sequence. This function will then provide a matrix that is sequences x match type. The best match will be reported based on a pre-defined list of rankings.

Usage

full_search(
  guide.seq,
  seqs,
  group.name = "siRNA",
  indel.bool = FALSE,
  allow_wobbles = FALSE,
  seqs.to.search = c("Full", "18mer", "15mer", "mer8", "mer7m8", "mer7A1", "mer6"),
  res_format = c("counts", "iranges"),
  fixed = FALSE,
  nonseed.edit.distances = c(0:4),
  seed.edit.distances = c(0:1)
)

Arguments

guide.seq

Input guide sequence.

seqs

Biostrings::DNAStringSet of sequences to search.

group.name

Group name to use for labeling.

indel.bool

Boolean in long format (TRUE, FALSE) for whether to include indels.

allow_wobbles

If true, all positions for a G:U wobble are considered.

seqs.to.search

List of sequences to search. Options: "Full", "18mer", "15mer", "mer8", "mer7m8", "mer7A1", "mer6".

res_format

Either "iranges" or "counts".

fixed

Whether to use matching string symbols for search when ambiguous.

nonseed.edit.distances

An integer value indicating allowed edit distance for non-seed searches.

seed.edit.distances

An integer value indicating allowed edit distance for seed searches.

Value

A data.frame of sequence match counts.

Examples

if (FALSE) { # interactive()
print("TBD")
}