Get the target sequence for a given siRNA input sequence

Given a sequence greater than 8 bp oriented 5' -> 3' and a seed definition (or custom definition), this function will return an object containing target sequence information.

Built-in options: mer8, mer7A1, mer7m8, mer6, Full, and 18mer

Custom Definitions: Users can input a custom seed name, but must provide the start position (start.pos) and stop position (stop.pos) that define the range of the seed sequence.

Note: mer8 and mer7A1 force a U at position g1. This results in an A in the target sequence being searched. This is based on historical data and observations from miRNA biology. If you do not wish to enforce this, provide the start, stop, and custom name to use the original sequence.

Usage

get_seed(
  guide.seq,
  seed.name = "mer7m8",
  start.pos = NULL,
  stop.pos = NULL,
  allow_wobbles = FALSE
)

Arguments

guide.seq

A character string greater than 8 bp and oriented 5'-> 3'.

seed.name

The seed name of interest. Options: mer8, mer7A1, mer7m8, mer6, 18-mer, 15-mer, or Full. If not in the default list, the start.pos and stop.pos arguments will be used to define the seed.

start.pos

The start position for a custom seed definition. 1-based

stop.pos

The stop position for a custom seed definition.

allow_wobbles

If True, allow G:U wobbles by replacing U with Y.

Value

An object with the entries:

Guide: Input guide sequence. Input is expected to be RNA.

Seed.Name: The seed name.

Seed.Seq.RNA: The seed sequence as a RNAString

Seed.Seq.DNA: The seed sequence as a DNAString

Target.Seq: The target DNA sequence based on the reverse complement of the Seed.Seq.DNA

Examples

# Example Ttr from Schlegel et al. 2022
guide.seq = "UUAUAGAGCAAGAACACUGUUUU"

# Get seed match
seed.seq = get_seed(guide.seq = guide.seq, seed.name = "mer7m8")