Load species-specific annotations from AnnotationHub

Use AnnotationHub::AnnotationHub() to load species-specific rtracklayer::TwoBitFile DNA sequence and ensembldb::EnsDb object. This function currently works for human, rat, and mouse.

The specific versions and builds used for each reference are further defined in articles(SeedMatchR_Annotation_Databases).

Usage

load_species_anno_db(
  reference.name = c("hg38", "hg38-old", "mm39", "mm10", "rnor6", "rnor7", "rnor7.113"),
  filter = NULL
)

Arguments

reference.name

Reference build name. Options: hg38, rnor6, rnor7, mm39, mm10

filter

An AnnotationFilter::AnnotationFilterList used to refine annotations.

Value

The function will return:

$gtf: A GenomicRanges::GRanges object containing the GTF information

$dna: A rtracklayer::TwoBitFile with reference DNA sequence

$txdb: A ensembldb::EnsDb object with transcriptome information

Examples

if (FALSE) { # interactive()
anno.db = SeedMatchR::load_species_anno_db("hg38")
}