Plot matches as alignments against target sequence feature

Plot matches as alignments against target sequence feature

Usage

plotAlignment(
  match.info,
  guide.seq,
  seqs,
  aln.name = "alignment",
  left.pad = 0,
  right.pad = 0,
  seq.start.pos = 2,
  seq.stop.pos = 8,
  output.type = c("data.frame", "ggplot")
)

Arguments

match.info

A row from the data.frame of matches produced by SeedMatchR

guide.seq

The full siRNA guide sequence

seqs

The feature sequences used in scanning

aln.name

The column name for the data.frame reporting alignments

left.pad

The number of nucleotides to the left to show on the feature sequence beyond the length of the guide sequence.

right.pad

The number of nucleotides to the right to show on the feature sequence beyond the length of the guide sequence.

seq.start.pos

The seed definition start position

seq.stop.pos

The seed definition stop position

output.type

Either data.frame or ggplot

Value

A data.frame representing the input sequences and alignment between them.