Get transcripts features and feature sequences

This function is used to get the genomic features of interest and the DNA sequences associated with them. This function takes advantage of the GenomicFeatures package functions threeUTRsByTranscript, fiveUTRsByTranscript, exonsBy, intronsByTranscript, and cdsBy. These functions are used to generate the features given an input tx.db object. A 2bit dna input is also required for extracting features sequences.

The output of the this function is:

$db: the feature GRanges object

$seqs: DNAStringSet of sequences associated to those features

Usage

get_feature_seqs(tx.db, dna, feature.type = "3UTR")

Arguments

tx.db

A tx.db object

dna

A 2bit dna sequence

feature.type

The type of feature to return. Options: 3UTR, 5UTR, exons, introns, cds

Value

list containing the feature db object and the feature sequences

Examples

if (FALSE) { # interactive()
anno.db = load_species_anno_db("human")

features = get_feature_seqs(anno.db$tx.db, anno.db$dna)
}